# services/core/message-hub/src/main/resources/client-api/aiassistant-openapi.yaml
openapi: 3.0.0
info:
  title: HoorayTalk AI Assistant API
  description: This is the HoorayTalk AI Assistant API specification for internal use.
  license:
    name: MIT License (MIT)
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
  - url: https://openai-assistant/v1
tags:
  - name: chat
    description: Chat operations
  - name: health
    description: Health check operations
paths:
  /api/v1/chat:
    post:
      tags:
        - chat
      summary: Generate a response to a message using OpenAI LLM
      description: Generate a response to a message using OpenAI LLM and send it to the target chat(s).
      operationId: generateResponse
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompleteChatRequest'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            text/plain:
              schema:
                type: string
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '500':
          description: Internal server error
      security:
        - IntraServiceApiKeyAuth: [ ]
  /api/v1/health:
    get:
      tags:
        - health
      summary: Service health check
      description: Service health check endpoint to verify the service is running
      operationId: getHealthStatus
      responses:
        '200':
          description: Service is healthy
        '401':
          description: Unauthorized
components:
  schemas:
    CompleteChatRequest:
      type: object
      required:
        - assistant_id
        - systemMessage
        - userMessage
      properties:
        assistant_id:
          type: string
          description: Unique identifier of the bot
        systemMessage:
          type: string
          description:
            The system message for the AI to generate the response.
        userMessage:
          type: string
          description:
            The user message for the AI to generate the response.
  securitySchemes:
    IntraServiceApiKeyAuth:
      type: apiKey
      name: API-KEY
      in: header